Cleanups - Added missing settings in .csproj files (Rootnamespace and ImplicitUsings) - Removed unused using statements - Renamed Backend.EmailIngestionproject to Backend.Mpa.EmailIngestion - Created new Backend.Mpa.Core project, and moved some services and handler to the new project -Moved some types from Backend.WebApi project to Backend.Core project
