+++ Backend.Backup/Managers/TenantBackupManager.cs
@@ -2,10 +2,8 @@ using System.Collections.Concurrent;
 using Backend.Backup.Providers;
 using Backend.Backup.Services;
 using Backend.Core;
-using Backend.Core.Infrastructure;
 using Backend.DbModel.Database;
 using Backend.DbModel.Database.EntityModels;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
