+++ Backend.Backup/Services/BackupService.cs
@@ -1,9 +1,7 @@
 using System.Text;
 using Backend.Backup.Providers;
 using Backend.Core;
-using Backend.Core.Infrastructure;
 using Backend.Core.Providers;
-using Backend.Crypto;
 using Backend.DbModel.Database.EntityModels;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json;
