+++ Backend.Crypto/Backend.Crypto.csproj
@@ -4,6 +4,7 @@
 		<TargetFramework>net10.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
+		<RootNamespace>Backend.Crypto</RootNamespace>
 	</PropertyGroup>
 
 </Project>
