+++ Backend.DbModel/Database/EFExtensions.cs
@@ -1,4 +1,3 @@
-using System.Collections;
 using System.Linq.Expressions;
 
 namespace Backend.DbModel.Database;
