+++ Backend.DbModel/Database/EntityModels/Blob.cs
@@ -1,7 +1,6 @@
 
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using Microsoft.EntityFrameworkCore;
 
 namespace Backend.DbModel.Database.EntityModels;
