+++ Backend.DbModel/Database/EntityModels/Tag.cs
@@ -2,7 +2,6 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.EntityFrameworkCore;
-using Backend.DbModel.Database;
 
 namespace Backend.DbModel.Database.EntityModels;
