+++ Backend.DbModel/Database/MpaDbContextFactory.cs
@@ -1,4 +1,3 @@
-using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
 
 namespace Backend.DbModel.Database;
