+++ Backend.Mpa.Core/Backend.Mpa.Core.csproj
@@ -4,14 +4,8 @@
 		<TargetFramework>net10.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
-		<RootNamespace>Backend.EmailIngestion</RootNamespace>
+		<RootNamespace>Backend.Mpa.Core</RootNamespace>
 	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="MailKit" Version="4.7.0" />
-		<PackageReference Include="Microsoft.Extensions.Configuration" Version="10.0.1" />
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="10.0.2" />
-	</ItemGroup>
 
 	<ItemGroup>
 		<ProjectReference Include="../Backend.Core/Backend.Core.csproj" />
