+++ Backend.Mpa.EmailIngestion/Cqrs/ExternalAccountHandler.cs
@@ -1,8 +1,8 @@
 using Backend.Core.Authentication;
 using Backend.Core.Cqrs.Infrastructure;
-using Backend.EmailIngestion.Services;
+using Backend.Mpa.EmailIngestion.Services;
 
-namespace Backend.EmailIngestion.Cqrs;
+namespace Backend.Mpa.EmailIngestion.Cqrs;
 
 
 public class ExternalAccount
