+++ Backend.WebApi/Backend.WebApi.csproj
@@ -24,7 +24,8 @@
 		<ProjectReference Include="../Backend.DbModel/Backend.DbModel.csproj" />
 		<ProjectReference Include="../Backend.Core/Backend.Core.csproj" />
 		<ProjectReference Include="../Backend.Crypto/Backend.Crypto.csproj" />
-		<ProjectReference Include="../Backend.EmailIngestion/Backend.EmailIngestion.csproj" />
+		<ProjectReference Include="../Backend.Mpa.EmailIngestion/Backend.Mpa.EmailIngestion.csproj" />
+		<ProjectReference Include="../Backend.Mpa.Core/Backend.Mpa.Core.csproj" />
 		<ProjectReference Include="..\Backend.Backup\Backend.Backup.csproj" />
 	</ItemGroup>
