+++ Backend.WebApi/Controllers/BlobController.cs
@@ -1,5 +1,5 @@
 using Backend.Core;
-using Backend.WebApi.Services;
+using Backend.Mpa.Core.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
