+++ Backend.WebApi/Middleware/TenantHeaderFromQueryStringMiddleware.cs
@@ -1,6 +1,3 @@
-using System.Diagnostics;
-
-
 namespace Backend.WebApi.Middleware;
