+++ Backend.WebApi/Middleware/TenantHeaderFromStateJsonMiddleware.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics;
 using System.Text.Json;
